adicionando API estoque

+++ src/main/java/com/virtualshelfshopping/Virtual/Shelf/Shopping/entity/Estoque.java
@@ -8,17 +8,17 @@ import jakarta.persistence.*;
 public class Estoque {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     private int quantProduto;
 
     // Getters and setters
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
@@ -29,5 +29,4 @@ public class Estoque {
     public void setQuantProduto(int quantProduto) {
         this.quantProduto = quantProduto;
     }
-
 }
